requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
langchain-chroma
langgraph
openai
python-dotenv
chromadb
pandas
docling